requirements.txt
pip~=23.2.1
pytest==8.4.2
playwright==1.55.0
PyYAML==6.0.3